+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - context: add `Context::try_send`.
 - errors: add `TrySendError::map` and `RequestError::map`.
 
+### Fixed
+- init: do not start termination if the memory tracker fails to read files.
+
 ## [0.1.32] - 2021-12-21
 ### Added
 - stream: add `Stream::generate()` to generate a stream from a generator.